unix fixes

+++ crc32.cpp
@@ -89,7 +89,7 @@ static uint32_t crc32_tab[] = {
 };
 
 uint32_t
-crc32(uint32_t crc, const void *buf, size_t size)
+crc32(uint32_t crc, const void *buf, int size)
 {
 	const uint8_t *p;
