+++ crc32.h
@@ -1,6 +1,6 @@
 #ifndef _CRC32_H
 #define _CRC32_H
 #include <stdint.h>
-uint32_t crc32(uint32_t crc, const void *buf, size_t size);
+uint32_t crc32(uint32_t crc, const void *buf, int size);
 
 #endif
